requirements.txt
pydantic
instructor
moviepy
google-generativeai